kotibai-dj-main/requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
bs4==0.0.2
CacheControl==0.14.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
diff-match-patch==20241021
distro==1.9.0
Django==5.1.2
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-import-export==4.2.0
django-jazzmin==3.0.1
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-standardized-errors==0.14.1
drf-yasg==1.21.8
firebase-admin==6.5.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
inflection==0.5.1
jiter==0.7.0
kombu==5.4.2
loguru==0.7.2
lxml==5.3.0
msgpack==1.1.0
openai==0.28.0
packaging==24.1
payme-pkg==2.6.6
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytubefix==8.3.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
tablib==3.7.0
tiktoken==0.8.0
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
